cleanup nuklear history page code
