+++ app/walletcore/helper.go
@@ -15,11 +15,7 @@ const (
 	DefaultRequiredConfirmations = 2
 
 	// default number of transactions to return per call to Wallet.TransactionHistory()
-	TransactionHistoryCountPerPage = 20
-
-	StoreSeedWarningText = "Keep the seed in a safe place as you will NOT be able to restore your wallet without it. " +
-		"Please keep in mind that anyone who has access to the seed can also restore your wallet " +
-		"thereby giving them access to all your funds, so it is imperative that you keep it in a secure location."
+	TransactionHistoryCountPerPage = 25
 
 	ReceivingDecredHint = "Each time you request payment, a new address is generated to protect your privacy."
 )
